Refactor GCN example
