+++ examples/pytorch/gcn/model.py
@@ -7,11 +7,10 @@ References:
 """
 import torch
 import torch.nn as nn
-from dgl.nn.pytorch import GraphConv
 
 class GCN(nn.Module):
     def __init__(self,
-                 g,
+                 g, layer,
                  in_feats,
                  n_hidden,
                  n_classes,
@@ -22,12 +21,12 @@ class GCN(nn.Module):
         self.g = g
         self.layers = nn.ModuleList()
         # input layer
-        self.layers.append(GraphConv(in_feats, n_hidden, activation=activation))
+        self.layers.append(layer(in_feats, n_hidden, activation=activation))
         # hidden layers
-        for i in range(n_layers - 1):
-            self.layers.append(GraphConv(n_hidden, n_hidden, activation=activation))
+        for i in range(n_layers):
+            self.layers.append(layer(n_hidden, n_hidden, activation=activation))
         # output layer
-        self.layers.append(GraphConv(n_hidden, n_classes))
+        self.layers.append(layer(n_hidden, n_classes, activation=None))
         self.dropout = nn.Dropout(p=dropout)
 
     def forward(self, features):
